Implementation of global exception, validations in user control and review of functionalities

+++ mvsc-users/src/main/java/org/dev/library/msvc/users/dev/Services/UsersService.java
@@ -1,5 +1,7 @@
 package org.dev.library.msvc.users.dev.Services;
 
+import org.dev.library.msvc.users.dev.Exceptions.UserNotFoundException;
+import org.dev.library.msvc.users.dev.Models.UserResponse;
 import org.dev.library.msvc.users.dev.Models.UsersModel;
 import org.dev.library.msvc.users.dev.Repositories.UsersRepositoryCustom;
 import org.dev.library.msvc.users.dev.Repositories.UsersRepository;
@@ -39,22 +41,32 @@ public class UsersService implements UsersRepositoryCustom {
 
     @Override
     @Transactional()
-    public UsersModel UpdateUser(UsersModel usuario) {
-        UsersModel existingUser  = usersRepository.findById(usuario.getId()).orElseThrow();
-        existingUser .setFirstname(usuario.getFirstname());
-        existingUser .setLastname(usuario.getLastname());
-        existingUser .setEmail(usuario.getEmail());
-        existingUser .setUsername(usuario.getUsername());
-        existingUser .setPhoneNumber(usuario.getPhoneNumber());
-        existingUser .setAddress(usuario.getAddress());
-        existingUser .setPassword(usuario.getPassword());
-        existingUser .setBirth(usuario.getBirth());
-        return usersRepository.save(existingUser );
+    public UsersModel UpdateUser(Long id, UsersModel usuario) {
+        Optional<UsersModel> existingUser  = usersRepository.findById(id);
+        if(existingUser.isPresent()) {
+            UsersModel existingUserModel = existingUser.get();
+            existingUserModel.setFirstname(usuario.getFirstname());
+            existingUserModel.setLastname(usuario.getLastname());
+            existingUserModel.setEmail(usuario.getEmail());
+            existingUserModel.setUsername(usuario.getUsername());
+            existingUserModel.setPhonenumber(usuario.getPhonenumber());
+            existingUserModel.setAddress(usuario.getAddress());
+            existingUserModel.setPassword(usuario.getPassword());
+            existingUserModel.setBirth(usuario.getBirth());
+            return usersRepository.save(existingUserModel);
+        }
+        throw new UserNotFoundException("User not found with id: " + id);
     }
 
     @Override
     @Transactional()
-    public void DeleteUser(Long id) {
-        usersRepository.deleteById(id);
+    public UserResponse DeleteByUser(Long id) {
+        Optional<UsersModel> optionalUser = usersRepository.findById(id);
+        if(optionalUser.isPresent()) {
+            usersRepository.deleteById(id);
+            return new UserResponse("User deleted successfully.", true);
+        } else {
+            throw new UserNotFoundException("User not found with id: " + id);
+        }
     }
 }
